Fix error of go_back trigger
The `on_enter_asleep` method needs `event` parameter but there is no any
parameter given when we directly trigger the function.

To fix the transition, I add `event` as the parameter into the `go_back`
function. Although it could be useless, it can satisfy the need of
`on_enter_asleep` function.